code refactored.

+++ src/main/java/org/aviansoundscape/training/preprocessing/FileProcessor.java
@@ -2,5 +2,5 @@ package org.aviansoundscape.training.preprocessing;
 
 public interface FileProcessor {
 
-    float[] processFile(String filePath);
+    double[] processFile(String filePath);
 }
